==> 8. COVID-19 - differential expression analysis/5. Correlate DE with classification patterns.ipynb <==
import pyreadr
import pickle
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

Define paths:

# this is the dataframe resulting from the analysis of individual classification trends:
path_classif_df = 'C:/Users/SG/Sex-bias-cell-type-classification-scRNAseq/7. Disease sets - analysis of classification results/Analysis_results/df_covid_classification_trend.pickle'

#### Import data

Import the DS results:

covid_result = pyreadr.read_r("final_DS_results_covid.rds")
df_DS_covid = covid_result[None]
df_DS_covid

Misclassification results

with open(path_classif_df, 'rb') as file:
    df_classification_trend_covid = pickle.load(file)
df_classification_trend_covid.sort_values('Classification trend')

#### Combine results

for i in df_DS_covid.index:
    if i in list(df_classification_trend_covid["Cell type"]):
        df_DS_covid.loc[i, "Classification trend"] = df_classification_trend_covid.loc[df_classification_trend_covid['Cell type'] == i,'Classification trend'].iloc[0]
    else:
        print(f'{i} not in misclassification analysis')

# round %DS for readability
def round_numeric(val, decimals=4):
    try:
        return round(float(val), decimals)
    except (ValueError, TypeError):
        return val
df_DS_covid['%DS'] = df_DS_covid['%DS'].apply(round_numeric)

# replace NaNs
df_DS_covid.replace('NaN', '-', inplace=True)
df_DS_covid.replace(float('nan'), '-', inplace=True)

# reformat table
df_DS_covid.reset_index(inplace = True)
df_DS_covid.rename(columns={'rownames': 'Cell type'}, inplace=True)
df_DS_covid = df_DS_covid[['Cell type', '#DS', '%DS', 'Classification trend']]

# make #DS an int for readability
def toint_numeric(val):
    try:
        return int(val)
    except (ValueError, TypeError):
        return val
df_DS_covid['#DS'] = df_DS_covid['#DS'].apply(toint_numeric)

# show table
df_DS_covid.head(60)

#### Rank sum test to check if distinct cell types are more DE

from scipy.stats import mannwhitneyu

df_totest = df_DS_covid[(df_DS_covid["#DS"] != '-') & (df_DS_covid["Classification trend"] != 'Inconclusive')]
df_totest['%DS'] = pd.to_numeric(df_totest['%DS'])

distinct = df_totest.loc[df_totest["Classification trend"] == 'Distinct','%DS']
nondistinct = df_totest.loc[df_totest["Classification trend"] == 'Non-distinct','%DS']

# perform the Mann-Whitney U test
stat, p_value = mannwhitneyu(distinct, nondistinct, alternative='greater')
stat, p_value

#### Investigate genes

Use Ensembl database to get gene name + chromosome number:

import requests

def get_gene_chromosome(ensembl_id):
    server = "https://rest.ensembl.org"
    ext = f"/lookup/id/{ensembl_id}?content-type=application/json"
    response = requests.get(f"{server}{ext}")
    if response.ok:
        gene_data = response.json()
        gene_name = gene_data.get('display_name', 'No data found')
        chromosome = gene_data.get('seq_region_name', 'No data found')
        return gene_name, chromosome
    else:
        return "Error fetching data", "Error fetching data"

pdc = pd.read_csv("result_Plasmacytoid DCs.csv")
pdc = pdc[["male.Plasmacytoid.DCs.gene", "male.Plasmacytoid.DCs.logFC", "male.Plasmacytoid.DCs.p_adj.loc"]]
pdc = pdc[pdc['male.Plasmacytoid.DCs.p_adj.loc'] < 0.05] 

gene_ids = list(pdc["male.Plasmacytoid.DCs.gene"])
gene_info = {gene_id: get_gene_chromosome(gene_id) for gene_id in gene_ids}

pdc_genes = pd.DataFrame.from_dict(gene_info, orient = 'index', columns=['Gene Name', 'Chromosome'])
pdc_genes.index.name = 'Gene ID'
pdc_genes.reset_index(inplace=True)

pdc_genes.sort_values('Chromosome')

len(pdc_genes[~pdc_genes["Chromosome"].isin(['MT', 'X', 'Y'])])

len(pdc_genes[pdc_genes["Chromosome"] == 'Y'])
